# Refactor: base GUI table on QStandardItemModel