--- a/damnit/backend/user_variables.py
+++ b/damnit/backend/user_variables.py
@@ -1,13 +1,9 @@
 from dataclasses import dataclass
 from typing import Optional
 
-import pandas as pd
-
 
 class ValueType:
     # These are intended to be overridden in subclasses.
-    type_instance = None
-
     type_name = None
 
     description = None
@@ -18,30 +14,15 @@ class ValueType:
         return self.type_name
 
     @classmethod
-    def unwrap(cls, x):
-        if len(x) == 1:
-            return x.to_numpy().item()
-        return list(x)
+    def parse(cls, input: str):
+        return input
 
     @classmethod
-    def convert(cls, data, unwrap=False):
-        is_string = isinstance(data, str)
-        is_sequence = hasattr(data, "__len__") and not is_string
-
-        if not is_sequence:
-            data = [data]
-
-        if len(data) > 0:
-            res = pd.Series(data).convert_dtypes().astype(cls.type_instance)
-        else:
-            res = pd.Series([], dtype=cls.type_instance)
-
-        return cls.unwrap(res) if unwrap else res
+    def from_db_value(cls, value):
+        return value
 
 
 class BooleanValueType(ValueType):
-    type_instance = pd.BooleanDtype()
-
     type_name = "boolean"
 
     description = "A value type that can be used to denote truth values."
@@ -68,53 +49,43 @@ class BooleanValueType(ValueType):
                 f"Value \"{to_convert}\" matches {'more than one' if n_matches > 0 else 'none'} of the allowed ones ({', '.join(valid_strings)})")
 
     @classmethod
-    def convert(cls, data, unwrap=False):
-        is_string = isinstance(data, str)
-        is_sequence = hasattr(data, "__len__") and not is_string
+    def parse(cls, input: str):
+        try:
+            return cls._valid_values[input.lower()]
+        except KeyError:
+            raise ValueError("{input!r} not one of true/false/1/0/yes/no")
 
-        if not is_sequence:
-            data = [data]
-
-        to_convert = pd.Series(data).convert_dtypes()
-        res = None
-
-        if len(data) > 0:
-            if to_convert.dtype == "object":
-                raise ValueError("The input array contains mixed object types")
-            elif to_convert.dtype == "string":
-                valid_strings = pd.Series(cls._valid_values.keys(), dtype="string")
-                res = to_convert.map(lambda x: cls._map_strings_to_values(x, valid_strings))
-            else:
-                res = to_convert.astype(cls.type_instance)
-        else:
-            res = pd.Series([], dtype=cls.type_instance)
-
-        return cls.unwrap(res) if unwrap else res
+    @classmethod
+    def from_db_value(cls, value):
+        if value is None:
+            return None
+        return bool(value)
 
 
 class IntegerValueType(ValueType):
-    type_instance = pd.Int32Dtype()
-
     type_name = "integer"
 
     description = "A value type that can be used to count whole number of elements or classes."
 
     examples = ["-7", "-2", "0", "10", "34"]
 
+    @classmethod
+    def parse(cls, input: str):
+        return int(input)
 
 class NumberValueType(ValueType):
-    type_instance = pd.Float32Dtype()
-
     type_name = "number"
 
     description = "A value type that can be used to represent decimal numbers."
 
     examples = ["-34.1e10", "-7.1", "-4", "0.0", "3.141592653589793", "85.4E7"]
 
+    @classmethod
+    def parse(cls, input: str):
+        return float(input)
+
 
 class StringValueType(ValueType):
-    type_instance = pd.StringDtype()
-
     type_name = "string"
 
     description = "A value type that can be used to represent text."