This is kind of necessary as a prelude to wiring the `Cell()` class up to the table in #221, but it's also something I've been working towards for a while.

Pandas is good for working with tables of numbers & doing calculations, but it's not a great match for what we want in a GUI: it doesn't really know about formatting or images, pulling out individual values to display in the GUI is not super efficient, and adding new rows is a bit awkward. Qt provides the [QStandardItemModel](https://doc.qt.io/qt-5/qstandarditemmodel.html) to represent rich tables in memory (as opposed to lazily loaded tables), and of course it fits neatly with GUI expectations, so this is what this PR uses instead of pandas.

I also discovered a neat piece of machinery to allow sorting (QSortFilterProxyModel), which means the order within the DamnitTableModel doesn't need to change, and we can always add new rows at the end. However, this does mean the indexes we get from the view have to be translated through the proxy model to find the relevant data.