@takluyver 2024-03-13
@tmichela asked me about performance, so I did some simple timings with a copy of the DAMNIT folder from p6616, which has 320 runs and 6742 variables stored. I measured from running the `amore-proto gui ...` command to when the window is drawn, in FastX on max-exfl-display.

- master: 19 s, 19 s, 18.5 s
- this branch: 9.7 s, 9.3 s, 9.6 s

So at least for this case, we've got a nice speedup.

When I've profiled in the past, the `DamnitTableModel.data()` method has taken a lot of time in total. This is what Qt calls to get information from the model, and even though it's quite fast, it's called so many times that small inefficiencies mount up.