@JamesWrigley 2024-03-13
BTW, this is quite an invasive change (and our tests are not as good as I thought they were) so I might suggest deploying it to the beta module for a couple experiments before deploying to the stable module.