@takluyver 2024-03-13
Thanks, that's a good idea. And this should be a case where people can freely switch between the stable & beta versions, so it's easy to fall back if something breaks.