@tmichela 2024-03-25
> If no one else spots anything then I'd say let's just deploy it to beta once this weeks experiments are over and see what happens 

I've also looked through the code and tested it without breaking it, so I'd agree to do that. 👍