@takluyver 2024-03-25
Thanks! Are you thinking we merge this and then deploy it to beta, or deploy it from the branch to give it some more testing before merge?